Remove remaining controllers.
